+++ CHANGELOG.md
@@ -4,11 +4,28 @@ Canonical reference for changes, improvements, and bugfixes for Boundary.
 
 ## Next
 
+## 0.7.5 (2022/02/17)
+
 ### New and Improved
 
 * cli: Update authentication examples to remove password flag and make
   subcommend selection a bit clearer
   ([PR](https://github.com/hashicorp/boundary/pull/1835))
+* Data Warehouse: Add addresses on plugin based hosts to the database warehouse.
+  3 new dimension tables have been added including `wh_network_address_group`
+  (which is now referenced by `wh_host_dimension`),
+  `wh_network_address_dimension`, and `wh_network_address_group_membership`.
+  ([PR](https://github.com/hashicorp/boundary/pull/1855))
+* ui: Add support for dynamic host catalog. AWS and Azure plugin-based CRUD operations.
+
+### Bug Fixes
+* targets: Specifying a plugin based host id when authorizing a session
+  now works. ([PR](https://github.com/hashicorp/boundary/pull/1853))
+* targets: DNS names are now properly parsed when selecting an endpoint
+  for authorizing a session.
+  ([PR](https://github.com/hashicorp/boundary/pull/1849))
+* hosts: Static hosts now include the host sets they are in.
+  ([PR](https://github.com/hashicorp/boundary/pull/1828))
 
 ## 0.7.4 (2022/01/18)